requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
pydantic==2.7.2
pydantic_core==2.18.3
pytest==8.2.1
pytest-check==2.3.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==70.0.0
typing_extensions==4.12.0
urllib3==2.2.1
wheel==0.43.0
